Workers take time to be trained
There is a queue of units to be trained. Each update, the units in the queue is delegated to available training halls. Training halls will spawn units and tell them to go to some location.

Fixed issue when converting mouse position to raycast. Only the rotation of the camera should be used when converting and not the whole view matrix which includes position and rotation.